Clarify notification sticky toggle checkbox label
When RES notification popups appear in the upper right of the page, they contain a checkbox labeled "Enabled".

As a new user it was unclear to me that this checkbox actually controls whether this type of notification will be shown again the next time it is triggered. 

Specifically I had enabled the NSFW filter and started seeing popups when filtering occurred. I assumed that the "Enabled" checkbox, if unchecked, would turn off the NSFW filtering. The text that appears in the popup further leads one to look elsewhere for a way to hide the notification ("To hide this message, disable the filteReddit notification.").

This pull request changes the label from "Enabled" to "Show next time". This should disambiguate the function of this checkbox.